Extract HtmlEditButton and htmlEditButtonOptions
